rename: SelectCardComponent -> RadioCard, type 'radiocard'
Reviewer feedback on #52: drop the redundant Component suffix from
the class/file names and lowercase the type string per repo convention.
