+++ src/templates/bootstrap/index.ts
@@ -1,11 +1,11 @@
 import checkmatrix from './checkmatrix';
-import selectCardComponent from './selectCardComponent';
+import radiocard from './radiocard';
 import tree from './tree';
 import treePartials from './tree/partials';
 
 export default {
   checkmatrix,
-  selectCardComponent,
+  radiocard,
   tree,
   ...treePartials,
 };
