+++ src/templates/bootstrap/radiocard/index.ts
